requirements.txt
streamlit>=1.34
requests>=2.32
python-dotenv>=1.0
pandas>=2.2
